Riddhi changes

+++ app/Http/Controllers/StripeController.php
@@ -12,43 +12,60 @@ class StripeController extends Controller
 {
     public function stripe(Request $request)
     {
-        
-        
+       // dd($request->all());
        $checkpayment =Order::where('customer_id',$request->customerid)->where(['active_status'=>1])->first();
+
         if($checkpayment == null)
         {
 
             if($request->paymentmethod == "offline")
             {    
-                $tempcustomer = TempCustomer::where('id',$request->customerid)->first();
+                if($request->usertype =='freetopro')
+                {
+                    $customer = Customer::find($request->customerid);
+                    $customer->access_type= 'requestforpaid';
+                    $customer->save();
 
-                $customer = new Customer();
-                $customer->fname=$tempcustomer->fname;
-                $customer->lname=$tempcustomer->lname;
-                $customer->job_title=$tempcustomer->job_title;
-                $customer->phone=$tempcustomer->phone;
-                $customer->email=$tempcustomer->email;
-                $customer->bussiness_name=$tempcustomer->bussiness_name;
-                $customer->business_wider_group=$tempcustomer->business_wider_group;
-                $customer->additional_details=$tempcustomer->additional_details;
-                $customer->additional_user_no =$tempcustomer->additional_user_no;
-                $customer->remainadditional_user=$tempcustomer->additional_user_no;
-                $customer->gst=$tempcustomer->gst;
-                $customer->access_type= 'requestforpaid';
-                $customer->save();
+                    $order = new Order();
+                    $order->customer_id=$customer->id;
+                    $order->membershipplans_id=$request->membershipid;
+                    $order->amount=$request->membershipprice;
+                    $order->currency='USD';
+                    $order->payment_method=$request->paymentmethod;
+                    $order->save();
+                    return redirect()->route('offlinepaymentsuccess');
+                }else{
 
-                $order = new Order();
-                $order->customer_id=$customer->id;
-                $order->membershipplans_id=$request->membershipid;
-                $order->amount=$request->membershipprice;
-                $order->currency='USD';
-                $order->payment_method=$request->paymentmethod;
-                $order->save();
+                    $tempcustomer = TempCustomer::where('id',$request->customerid)->first();
+                    $customer = new Customer();
+                    $customer->fname=$tempcustomer->fname;
+                    $customer->lname=$tempcustomer->lname;
+                    $customer->job_title=$tempcustomer->job_title;
+                    $customer->phone=$tempcustomer->phone;
+                    $customer->email=$tempcustomer->email;
+                    $customer->bussiness_name=$tempcustomer->bussiness_name;
+                    $customer->business_wider_group=$tempcustomer->business_wider_group;
+                    $customer->additional_details=$tempcustomer->additional_details;
+                    $customer->additional_user_no =$tempcustomer->additional_user_no;
+                    $customer->remainadditional_user=$tempcustomer->additional_user_no;
+                    $customer->gst=$tempcustomer->gst;
+                    $customer->access_type= 'requestforpaid';
+                    $customer->save();
 
-                $tempcustomer = TempCustomer::where('id',$request->customerid)->delete();
-                session()->forget('customer'); //forgotp old session
-                $session= Session::put('customer', $customer->id); //genrate new session
-                return redirect()->route('offlinepaymentsuccess');
+                    $order = new Order();
+                    $order->customer_id=$customer->id;
+                    $order->membershipplans_id=$request->membershipid;
+                    $order->amount=$request->membershipprice;
+                    $order->currency='USD';
+                    $order->payment_method=$request->paymentmethod;
+                    $order->save();
+
+                    $tempcustomer = TempCustomer::where('id',$request->customerid)->delete();
+                    session()->forget('customer'); //forgotp old session
+                    $session= Session::put('customer', $customer->id); //genrate new session
+                    return redirect()->route('offlinepaymentsuccess');
+                }
+                
             }
             else
             {
@@ -63,7 +80,24 @@ class StripeController extends Controller
 
                 if($response->status == 'succeeded')
                 {      
-                    // dd($request->all());
+                    if($request->usertype =='freetopro')
+                    {
+                        $customer = Customer::find($request->customerid);
+                        $customer->access_type= 'requestforpaid';
+                        $customer->save();
+
+                        $order = new Order();
+                        $order->customer_id = $customer->id;
+                        $order->membershipplans_id = $request->membershipid;
+                        $order->amount = $request->membershipamount;
+                        $order->currency = 'USD';
+                        $order->payment_method = $request->paymentmethod;
+                        $order->payment_id = $response->id;
+                        $order->receipt_url = $response->receipt_url;
+                        $order->status = 1;
+                        $order->save();
+                        return redirect()->route('offlinepaymentsuccess');
+                    }else{
                     $tempcustomer = TempCustomer::where('id',$request->customerid)->first();
 
                     $customer = new Customer();
@@ -96,7 +130,8 @@ class StripeController extends Controller
                     $tempcustomer = TempCustomer::where('id',$request->customerid)->delete();
                     session()->forget('customer'); //forgotp old session
                     $session= Session::put('customer', $customer->id); //genrate new session
-                    return redirect()->route('offlinepaymentsuccess');    
+                    return redirect()->route('offlinepaymentsuccess');  
+                    }  
                 }else{
 
                       $tempcustomer = TempCustomer::where('id',$request->customerid)->delete();
